requirements.txt
langchain>=0.1.0
langchain-community>=0.0.10
langchain-core>=0.1.10
langchain-huggingface>=0.0.10
chromadb>=0.4.18
sentence-transformers>=2.2.2
pypdf>=3.17.0
gradio>=4.0.0
python-dotenv>=1.0.0